AP_RangeFinder: initial fixes for uLanding Gaussian filter
Fixes some type definitions from int to float.

Fixes the kernel initialization to not add to the kernelsum for the
middle index of the kernel array.

Fixes input_array padding to shift the end of the array by wl.

Fixes the indexing on input_array during application of the kernel.

Adds a debug printout option with a compile time flag.